Collapse the flattening one level: Branch card is also branch header
Collapse the logic of the branch header component into the branch header.
