+++ apps/desktop/src/components/v3/StackDraft.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
 	import BranchCard from '$components/v3/BranchCard.svelte';
-	import BranchHeader from '$components/v3/BranchHeader.svelte';
 	import CommitGoesHere from '$components/v3/CommitGoesHere.svelte';
 	import { StackService } from '$lib/stacks/stackService.svelte';
 	import { UiState } from '$lib/state/uiState.svelte';
@@ -34,18 +33,14 @@
 </script>
 
 <div data-testid={TestId.StackDraft} class="stack-draft">
-	<BranchCard type="draft-branch" {projectId} {branchName}>
-		{#snippet header()}
-			<BranchHeader
-				type="draft-branch"
-				{branchName}
-				{projectId}
-				iconName="branch-local"
-				readonly={false}
-				lineColor="var(--clr-commit-local)"
-			/>
-		{/snippet}
-	</BranchCard>
+	<BranchCard
+		type="draft-branch"
+		{projectId}
+		{branchName}
+		iconName="branch-local"
+		readonly={false}
+		lineColor="var(--clr-commit-local)"
+	/>
 	<CommitGoesHere selected draft />
 </div>
